Nouns.java
Write a program that takes 10 nouns as strings and form their plural on the basis of
these rules:
a. If noun ends in ‘y’, remove the ‘y’ and add ‘ies’
b. If noun ends in ‘s’, “ch”, or “sh” add ‘es’
c. In all other cases, just add ‘s’.
Print each noun and their plurals.